# Build a type encoder caching system for the `form` package