@brandur-stripe 2017-09-16
Also worth noting that way more optimization work could be done here — I didn't even pull out a profiler for this one. I think it's worth pulling in the 80% win first though, and then moving onto to coaxing out the last 20%.