@brandur-stripe 2017-09-17
> This looks good to me but I'll admit it goes a bit beyond what I could grasp just reviewing so I'm not confident I would catch an edge-case or something you missed.

Yeah I totally get it. I think in this sort of case we need to be relying heavily on the test suite, which did catch quite a few problems as I was writing it.

Thanks for taking a look!