@brandur-stripe 2017-09-17
Released all recent changes as 26.0.0.